cleaned up findPoles, runs once, angle limits are trouble.

+++ biomech_flexsea-projects/MIT_DLEG/inc/actuator_functions.h
@@ -61,7 +61,7 @@ void  setMotorTorque(struct act_s *actx, float tor_d);
 void  packRigidVars(struct act_s *actx);
 
 void  setMotorTorqueOpenLoop(struct act_s *actx, float tau_des);
-float frequencySweep(float omega, uint32_t t);
+float frequencySweep(float omega, float t);
 
 
 //Smoothing
